Move Embla Carousel to tile Editor folder
